requirements.txt
Django==2.2.4
django-address==0.2.1
django-autocomplete-light==3.4.1
django-crispy-forms==1.7.2
# django-jet==1.0.8
django-model-utils==3.2.0
django-annoying==0.10.4
django_extensions>=2.2.1
django-tempus-dominus==5.1.2.8
psycopg2>=2.8.3
nhpgd>=0.204
phonenumbers==8.10.19
psycopg2-binary>=2.8.3
sisa>=0.918
pycodestyle==2.5.0
pyexcel>=0.5.15
pyexcel_xls>=0.5.8
cie>=0.204
fastkml>=0.11
django_tinymce>=2.8.0
oss-ar>=0.131
